refactor: ai 파일 수정
